fix(purchase-returns): record returns as purchase returns, with location, items and the right date
Tenant report (boi-enterprises): an approved purchase return showed in the
item's stock history as SOLD / "Order 2586cf93", its date came out a day
early, the return never recorded which outlet/location the goods left from,
and the returned items were not shown anywhere.

- Approval now takes the goods out as a "purchase_return" stock adjustment
  (new reason) referencing the return number, instead of going through the
  sales consumption path. The ledger labels it Purchase Return with the
  return number and supplier. The reason is excluded from stock.adjusted GL
  events (treasury values the return through the vendor credit note) and
  matches the reason treasury's eTIMS subscriber already skips, so KRA sees
  the return once (stockIO 12).
- Approval is claimed atomically (pending to approved in one guarded update),
  so a double click can never take stock out twice; a second approve is 409.
  Every SKU is resolved before anything moves; any stock-out failure is
  returned to the approver as stock_warnings.
- Returns capture warehouse_id (defaults to the operating outlet's warehouse,
  same resolution as goods receipts). Create saves header and lines in one
  transaction and validates items/quantities (decimal quantities now allowed).
  The date is a calendar day and cannot be in the future.
- List/Get return supplier and location names, date_returned_day, item count,
  items summary and full lines; filters by supplier, warehouse, outlet and
  return number. The PDF shows the location.
- Legacy returns approved through the old consumption path are reclassified
  at read time in the stock ledger and excluded from ingredient utilization.
- The approved event now carries the return date, supplier name, reason,
  location and per-item names/sub-totals for an itemised treasury credit note.
- Schema: purchase_returns.warehouse_id, purchase_return_lines.quantity as
  double precision, stock_adjustments reason purchase_return. Swagger
  regenerated.

+++ internal/ent/purchasereturnline/purchasereturnline.go
@@ -66,7 +66,7 @@ func ValidColumn(column string) bool {
 
 var (
 	// DefaultQuantity holds the default value on creation for the "quantity" field.
-	DefaultQuantity int
+	DefaultQuantity float64
 	// DefaultSubTotal holds the default value on creation for the "sub_total" field.
 	DefaultSubTotal float64
 	// DefaultCreatedAt holds the default value on creation for the "created_at" field.
